pyproject.toml
[project]
name = "precinct"
version = "0.1.0"
description = "Democratic precinct analysis and campaign optimization tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.12.0",
    "black>=22.0.0",
    "click>=8.0.0",
    "dash>=2.14.0",
    "fiona>=1.8.0",
    "flake8>=5.0.0",
    "flask>=3.1.2",
    "flask-admin>=1.6.1",
    "flask-limiter>=4.0.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "geoalchemy2>=0.14.0",
    "isort>=5.10.0",
    "matplotlib>=3.10.7",
    "mypy>=0.991",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "plotly>=5.17.0",
    "psycopg2-binary>=2.9.0",
    "pyproj>=3.4.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "rtree>=1.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shapely>=2.0.0",
    "sqlalchemy>=2.0.0",
    "toml>=0.10.2",
    "urllib3>=2.0.0",
    "waitress>=3.0.2",
]

[dependency-groups]
dev = [
    "pytest-flask>=1.3.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
